Fix W4-02-owned ruff and format failures
Clear the static-check debt that originated in the shadow slice: Sequence
import path, redundant int() cast in the noise-mapped test helper, and
action_adapter raise formatting. Behavior is unchanged.

+++ src/lap/policies/cover_policy_shadow_test.py
@@ -114,7 +114,7 @@ class NoiseMappedCandidateGenerator:
         noise_arr = np.asarray(noise, dtype=np.float64)
         if noise_arr.ndim == 2:
             noise_arr = noise_arr[None, ...]
-        indices = [int(round(float(sample.reshape(-1)[0]))) % len(self._catalog) for sample in noise_arr]
+        indices = [round(float(sample.reshape(-1)[0])) % len(self._catalog) for sample in noise_arr]
         return self._catalog[indices].copy()
